Add config upload command
