+++ data_safe_haven/commands/config.py
@@ -25,3 +25,15 @@ def template(
             outfile.write(config.to_yaml())
     else:
         print(config.to_yaml())
+
+
+@config_command_group.command()
+def upload(
+    file: Annotated[Path, typer.Argument(help="Path to configuration file")]
+) -> None:
+    """Upload a configuration to the Data Safe Haven context"""
+    context = ContextSettings.from_file().context
+    with open(file) as config_file:
+        config_yaml = config_file.read()
+    config = Config.from_yaml(context, config_yaml)
+    config.upload()
